Split dynamic table stories into own file
